wip 4

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aantonio <aantonio@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 15:10:37 by aantonio          #+#    #+#             */
-/*   Updated: 2023/03/13 22:45:42 by aantonio         ###   ########.fr       */
+/*   Updated: 2023/03/14 15:44:55 by aantonio         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 # include <stddef.h>
 
 int		ft_printf(const char *str, ...);
-size_t	ft_put_ull(unsigned long nbr, char *base);
+unsigned long long	ft_put_ull(unsigned long nbr, char *base);
 
 #endif
